+++ src/main/java/net/thumbtack/school/notes/service/impl/SectionServiceImpl.java
@@ -20,6 +20,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
+/**
+ * Service to work with sections of notes on the server
+ */
 @Slf4j
 @RequiredArgsConstructor
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
@@ -29,6 +32,13 @@ public class SectionServiceImpl implements SectionService {
     SessionDaoImpl sessionDao;
     UserDaoImpl userDao;
 
+    /**
+     * Method to create new note`s section on the server
+     *
+     * @param createRequest contains name of new section
+     * @param sessionId     user session token
+     * @return created section information
+     */
     @Override
     @Transactional
     public Section createSection(SectionRequest createRequest, String sessionId) throws NoteServerException {
@@ -39,9 +49,18 @@ public class SectionServiceImpl implements SectionService {
         return sectionDao.createSection(section);
     }
 
+    /**
+     * Method to rename existing section name
+     *
+     * @param renameRequest contains new section name
+     * @param sessionId     current user session token
+     * @param sectionId     identifier of section to be renamed
+     * @return created section information
+     * @throws NoteServerException if owner of current session is not the author of section with given identifier
+     */
     @Override
     @Transactional
-    public Section renameSection(SectionRequest renameRequest, String sessionId, long sectionId) throws NoteServerException {
+    public Section renameSection(SectionRequest renameRequest, String sessionId, int sectionId) throws NoteServerException {
         log.info("Trying to rename existed section");
         int userId = sessionDao.getSessionBySessionId(sessionId).getUserId();
         Section section = sectionDao.getSectionInfo(sectionId);
@@ -51,9 +70,18 @@ public class SectionServiceImpl implements SectionService {
         return sectionDao.renameSection(sectionId, section.getSectionName());
     }
 
+    /**
+     * Method to delete note`s section from the server
+     * User can only delete section whose author he is
+     * Admin can delete any section
+     *
+     * @param sessionId current user session token
+     * @param sectionId identifier of section to be deleted
+     * @throws NoteServerException if owner of current session is not the author of section with given identifier
+     */
     @Override
     @Transactional
-    public void deleteSection(String sessionId, long sectionId) throws NoteServerException {
+    public void deleteSection(String sessionId, int sectionId) throws NoteServerException {
         log.info("Trying to delete section");
         int userId = sessionDao.getSessionBySessionId(sessionId).getUserId();
         Section section = sectionDao.getSectionInfo(sectionId);
@@ -66,14 +94,27 @@ public class SectionServiceImpl implements SectionService {
         sectionDao.deleteSection(sectionId);
     }
 
+    /**
+     * Method to get existing section information
+     *
+     * @param sessionId current user session token
+     * @param sectionId identifier of section for which information is needed
+     * @return information about section with given identifier
+     */
     @Override
     @Transactional
-    public Section getSectionInfo(String sessionId, long sectionId) throws NoteServerException {
+    public Section getSectionInfo(String sessionId, int sectionId) throws NoteServerException {
         log.info("Trying to get information about section");
         sessionDao.getSessionBySessionId(sessionId);
         return sectionDao.getSectionInfo(sectionId);
     }
 
+    /**
+     * Method to get information about all sections on the server
+     *
+     * @param sessionId current user session token
+     * @return information about all existing sections on the server in list
+     */
     @Override
     @Transactional
     public List<Section> getAllSections(String sessionId) throws NoteServerException {
